fix: 인증 보안 개선 및 사용자 이름 표시

+++ backend/src/main/java/com/run4you/brand/dto/BrandUpdateCommissionRequest.java
@@ -3,11 +3,12 @@ package com.run4you.brand.dto;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 import lombok.Getter;
+import java.math.BigDecimal;
 
 @Getter
 public class BrandUpdateCommissionRequest {
 
     @NotNull(message = "수수료율을 입력해주세요.")
     @Positive(message = "수수료율은 0보다 커야 합니다.")
-    private Double commissionRate;
+    private BigDecimal commissionRate;
 }
